app/Http/Requests/Manager/Lessonディレクトリが肥大化しているための整理(yuki)
