+++ app/Http/Requests/Manager/Lesson/BulkDeleteRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Lesson;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class LessonUpdateTitleRequest extends FormRequest
+class BulkDeleteRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -21,7 +21,6 @@ class LessonUpdateTitleRequest extends FormRequest
         $this->merge([
             'course_id' => $this->route('course_id'),
             'chapter_id' => $this->route('chapter_id'),
-            'lesson_id' => $this->route('lesson_id'),
         ]);
     }
 
@@ -35,8 +34,8 @@ class LessonUpdateTitleRequest extends FormRequest
         return [
             'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
             'chapter_id' => ['required', 'integer', 'exists:chapters,id,deleted_at,NULL'],
-            'lesson_id' => ['required', 'integer', 'exists:lessons,id,deleted_at,NULL'],
-            'title' => ['required', 'string', 'max:50'],
+            'lessons' => ['required', 'array'],
+            'lessons.*' => ['required', 'integer', 'exists:lessons,id,deleted_at,NULL'],
         ];
     }
 }
