+++ app/Http/Requests/Manager/Lesson/DeleteAllRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Lesson;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class LessonDeleteRequest extends FormRequest
+class DeleteAllRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -16,12 +16,11 @@ class LessonDeleteRequest extends FormRequest
         return true;
     }
 
-    protected function prepareForValidation()
+    protected function prepareForValidation(): void
     {
         $this->merge([
             'course_id' => $this->route('course_id'),
             'chapter_id' => $this->route('chapter_id'),
-            'lesson_id' => $this->route('lesson_id'),
         ]);
     }
 
@@ -35,7 +34,6 @@ class LessonDeleteRequest extends FormRequest
         return [
             'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
             'chapter_id' => ['required', 'integer', 'exists:chapters,id,deleted_at,NULL'],
-            'lesson_id' => ['required', 'integer', 'exists:lessons,id,deleted_at,NULL'],
         ];
     }
 }
