+++ app/Http/Requests/Manager/Lesson/PutRequest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Lesson;
 
+use App\Rules\LessonStatusRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class LessonBulkDeleteRequest extends FormRequest
+class PutRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -21,6 +22,7 @@ class LessonBulkDeleteRequest extends FormRequest
         $this->merge([
             'course_id' => $this->route('course_id'),
             'chapter_id' => $this->route('chapter_id'),
+            'lesson_id' => $this->route('lesson_id'),
         ]);
     }
 
@@ -34,8 +36,11 @@ class LessonBulkDeleteRequest extends FormRequest
         return [
             'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
             'chapter_id' => ['required', 'integer', 'exists:chapters,id,deleted_at,NULL'],
-            'lessons' => ['required', 'array'],
-            'lessons.*' => ['required', 'integer', 'exists:lessons,id,deleted_at,NULL'],
+            'lesson_id' => ['required', 'integer', 'exists:lessons,id,deleted_at,NULL'],
+            'title' => ['required', 'string', 'max:50'],
+            'url' => ['required', 'string'],
+            'remarks' => ['nullable', 'string'],
+            'status' => ['required', 'string', new LessonStatusRule],
         ];
     }
 }
