+++ app/Http/Requests/Manager/Lesson/PutStatusRequest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Lesson;
 
+use App\Rules\LessonStatusRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class LessonsAllDeleteRequest extends FormRequest
+class PutStatusRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -16,7 +17,7 @@ class LessonsAllDeleteRequest extends FormRequest
         return true;
     }
 
-    protected function prepareForValidation(): void
+    protected function prepareForValidation()
     {
         $this->merge([
             'course_id' => $this->route('course_id'),
@@ -34,6 +35,9 @@ class LessonsAllDeleteRequest extends FormRequest
         return [
             'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
             'chapter_id' => ['required', 'integer', 'exists:chapters,id,deleted_at,NULL'],
+            'lessons' => ['required', 'array'],
+            'lessons.*' => ['required', 'integer', 'exists:lessons,id,deleted_at,NULL'],
+            'status' => ['required', 'string', new LessonStatusRule],
         ];
     }
 }
