+++ app/Http/Requests/Manager/Lesson/StoreRequest.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Lesson;
 
-use App\Rules\LessonStatusRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class LessonPutRequest extends FormRequest
+class StoreRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,7 +21,6 @@ class LessonPutRequest extends FormRequest
         $this->merge([
             'course_id' => $this->route('course_id'),
             'chapter_id' => $this->route('chapter_id'),
-            'lesson_id' => $this->route('lesson_id'),
         ]);
     }
 
@@ -36,11 +34,7 @@ class LessonPutRequest extends FormRequest
         return [
             'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
             'chapter_id' => ['required', 'integer', 'exists:chapters,id,deleted_at,NULL'],
-            'lesson_id' => ['required', 'integer', 'exists:lessons,id,deleted_at,NULL'],
             'title' => ['required', 'string', 'max:50'],
-            'url' => ['required', 'string'],
-            'remarks' => ['nullable', 'string'],
-            'status' => ['required', 'string', new LessonStatusRule],
         ];
     }
 }
