+++ app/Http/Requests/Manager/Lesson/UpdateTitleRequest.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Lesson;
 
-use App\Rules\LessonStatusRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class LessonPutStatusRequest extends FormRequest
+class UpdateTitleRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,6 +21,7 @@ class LessonPutStatusRequest extends FormRequest
         $this->merge([
             'course_id' => $this->route('course_id'),
             'chapter_id' => $this->route('chapter_id'),
+            'lesson_id' => $this->route('lesson_id'),
         ]);
     }
 
@@ -35,9 +35,8 @@ class LessonPutStatusRequest extends FormRequest
         return [
             'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
             'chapter_id' => ['required', 'integer', 'exists:chapters,id,deleted_at,NULL'],
-            'lessons' => ['required', 'array'],
-            'lessons.*' => ['required', 'integer', 'exists:lessons,id,deleted_at,NULL'],
-            'status' => ['required', 'string', new LessonStatusRule],
+            'lesson_id' => ['required', 'integer', 'exists:lessons,id,deleted_at,NULL'],
+            'title' => ['required', 'string', 'max:50'],
         ];
     }
 }
